# tools: speedup compilation of js2c output