--- a/node.gyp
+++ b/node.gyp
@@ -1192,6 +1192,9 @@
         [ 'node_shared_libuv=="false"', {
           'dependencies': [ 'deps/uv/uv.gyp:libuv#host' ],
         }],
+        [ 'OS in "linux mac"', {
+          'defines': ['NODE_JS2C_USE_STRING_LITERALS'],
+        }],
         [ 'debug_node=="true"', {
           'cflags!': [ '-O3' ],
           'cflags': [ '-g', '-O0' ],